Fixed kernel loading
The call to loadEverything actually does load everything, including the
bodies of methods from other libraries. Getting kernel to correctly load
the needed signatures from other libraries is difficult, and can be
delayed until we have a working compiler (especially since we might
change frontends entirely once kernel has strong-mode support built in).

Instead, just keep a list of libraries that were passed in on the
command line, and only generate code for these.

Also added a simple AST visitor to dump the AST to stdout, so that we
can actually tell that the compiler is doing something. debug_printer is
not meant to be anywhere near production level, but it's useful in these
early stages of development.

BUG=
R=jmesserly@google.com, stanm@google.com

Review URL: https://chromereviews.googleplex.com//455977013 .