Print usage information on bad arguments
dart2java.dart used to silently do nothing if no arguments were given.
This is not helpful if you don't know how to use it.

Now it prints the available commands if none are provided. If the
'compile' command is specified but no arguments are given to it,
a message to provide source files as arguments is printed, together with
the available options for the compile command.

BUG= https://b.corp.google.com/issues/29905956
R=andrewkrieger@google.com, jmesserly@google.com

Review URL: https://chromereviews.googleplex.com//465677013 .