Pull package:args/command_runner.dart into d2j
This is a preparation step for removing commands from the compiler. We
are removing comands, because we don't need them / there is a single
real command - compile. The only other command we inherited from DDC is
'help', which we can add as a command line option '--help'.

BUG=
R=jmesserly@google.com

Review URL: https://chromereviews.googleplex.com//466737014 .