Expose analyzer metadata on classes and members
This commit exposes metadata in Kernel classes and members. The metadata
is exposed as a list of analyzer DartObjects, which means that now the
kernel AST depends directly on analyzer. This is a temporary hack, meant
to get ddc-java running faster. It should be replaced once kernel adds
support for metadata.