Make methods in interceptor classes non-static.
This allows for virtual method calls of methods defined on interceptor
classes. For example, we can now run the following Dart code:

7.hashCode;

That works even though hashCode is defined on Object and not on int,
i.e., it is defined on JavaObject and not on JavaInteger.

BUG=
R=jmesserly@google.com

Review URL: https://chromereviews.googleplex.com//479727013 .