Implement DartList with generics.
Provides a Java implementation of Dart lists. Generic arguments are
stored as a field in ClassOrInterfaceType and DartList (which is a
wrapper for arrays). Static method invocations (e.g. calls to the
factories) have the type arguments as the first parameters. In generated
the Java code, (reified) generic types are represented as generic
java.lang.Class objects. This is likely going to change once we have
proper type representations.

BUG=
R=stanm@google.com

Review URL: https://chromereviews.googleplex.com//486787013 .