Implement Dart constructor semantics.
The body of a Dart constructor is part of a Java instance method. That
gives us precise control of when to call super, initialize fields, and
run the constructor body. Fields are no longer final in generated Java
code even if they are annotated as such, because they would have to be
initialized in a Java constructor.

BUG=
R=andrewkrieger@google.com

Review URL: https://chromereviews.googleplex.com//490637017 .