Pass runtime types into constructors.
It's not clear that we should always pass types into constructors (especially
for non-generic classes, it creates a lot of noise in the output). However,
since we want to be able to pass types in constructors sometimes (in order to
hoist type representations to constant variables), we at least want this
ability. We can add alternative constructors that are less noisy later.

BUG=
R=leafp@google.com

Review URL: https://chromereviews.googleplex.com//493167013 .