Hoist type expressions to static variables.
Allocate type expressions statically instead of creating them on
demand. For code that does type checks in inner loops, this results in
drastically fewer allocations. This also paves the way for hoisting
type representations (to reduce the number of type expression
evaluations that are needed).

Add a notion of "ClassState" to lib/src/java/type_system.dart. This
state currently holds the list of type expressions for every generated
Java class. The ClassState also takes responsibility for creating the
type-info fields, so the compiler no longer has to understand the
importance of the type info fields and their relative ordering.

The code changes are fairly small (only java_builder and type_system);
most of the bulk of this commit is from changes in codegen_expect.

BUG=
R=leafp@google.com, springerm@google.com

Review URL: https://chromereviews.googleplex.com//505967013 .