Fix name mangling of method invocations through supertype.
This commit does not yet handle generic specializations with
superclasses correctly. This will be done in the next commit. This
commit just make sure that the "specialization suffix" that is appended
to method names is correct. It does not yet append the class name of the
method.

BUG=
R=vsm@google.com

Review URL: https://codereview.chromium.org/2377573003 .